Make VirtualMachineRegister more extensible
In preparation for creating a JitBuilder extension for VirtualMachineRegister,
to facilitate the client API creation, the VirtualMachineRegister class needs
to become more like a fully extensible with an independent header.

Also change references to OMR::VirtualMachineRegister to TR::VirtualMachineRegister.
